// Frontend/Angular/src/app/clients/teacher/model/groupTableDto.ts
/**
 * Teacher api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AssignmentDto } from './assignmentDto';
import { GroupTableEntryDto } from './groupTableEntryDto';

export interface GroupTableDto { 
    assignments?: Array<AssignmentDto>;
    entries?: Array<GroupTableEntryDto>;
}