// Frontend/Angular/src/app/clients/teacher/model/studentAssignmentDto.ts
/**
 * Teacher api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AssignmentStatus } from './assignmentStatus';
import { UserListDto } from './userListDto';

export interface StudentAssignmentDto { 
    id?: number;
    assignmentId?: number;
    groupId?: number;
    studentId?: number;
    status?: AssignmentStatus;
    score?: number;
    fineScore?: number;
    submitted?: Date;
    lastStatusChange?: Date;
    completed?: Date;
    student?: UserListDto;
    groupName?: string;
    assignmentName?: string;
}